fix(credentialless): force clear certs on guest -> reg
in redesign use connection

Jira-Id: VPNAPPL-2979

+++ apps/ios/ProtonVPN/Scenes/Common/Services/iOSNetworkingDelegate.swift
@@ -13,6 +13,7 @@ import Foundation
 import GoLibs
 import LegacyCommon
 import ProtonCoreDataModel
+import ProtonCoreFeatureFlags
 import ProtonCoreForceUpgrade
 import ProtonCoreHumanVerification
 import ProtonCoreNetworking
@@ -52,9 +53,18 @@ final class iOSNetworkingDelegate: NetworkingDelegate {
     }
 
     func onGuestToAuthenticatedTransition() {
-        @Dependency(\.vpnAuthenticationStorage) var authenticationStorage
-        authenticationStorage.deleteKeys()
-        authenticationStorage.deleteCertificate()
+        if FeatureFlagsRepository.isRedesigniOSEnabled {
+            // with redesign push intent via bridge
+            @Dependency(\.connectionBridge) var bridge
+            Task {
+                await bridge.push(intent: .onSessionChange)
+            }
+        } else {
+            // pre-redesign delete manually
+            @Dependency(\.vpnAuthenticationStorage) var authenticationStorage
+            authenticationStorage.deleteKeys()
+            authenticationStorage.deleteCertificate()
+        }
         log.info("Cleared VPN authentication data during guest to authenticated transition", category: .net)
     }
 }
